SEO Tools: hide meta description when Yoast is active
Trying to save meta description for posts will fail if conflicting
SEO plugins are active on Jetpack site. We should hide the SEO
accordion in that case in order to prevent errors while saving.
